Add disabled test for emojis in comments.
It's not very clear why adding an emoji into a comment within a function causes the parser to get confused. I believe this may lie in the `skipToEndOfObject` function, but this will require some further investigation. The same issue does not appear to affect classes.

Whilst investigating, I took the opportunity to perform some cleanup around the codebase, including:
- Restricting the scope of functions defined in grammar files.
- Removing duplicated logic in `TokenSubStream`.
- Adding extra tests around the parsing of comments, particularly around emojis.
- General code style cleanup.