upgrade to scala 2.13.1
based on https://github.com/ShiftLeftSecurity/fuzzyc2cpg/pull/94 which
had to be reverted.

Some scalac options have been removed, others have been renamed. I
hope I kept the intentions as before.

Unfortunately we have a cyclic dependency with codepropertygraph, but
since it's only a test dependency I took the shortcut of publishing a
2.13 release for codepropertygraph which we use here. Once fuzzyc2cpg
is published for scala 2.13, we can release a normal cpg version which
we can depend on here.
Long term we should resolve the cyclic dependency.